improve the test
